at: actions: Do not track generated files
The files under gen/lib are generated by gnulib-tool and as generated
files they should not be tracked.
This would cause problems for a straight autoreconf, configure and make
as the files now do not exist after the clone for the actions, therefore
add a gnulib-tool --update to create the files.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>