docs: Create a reproducible archive for html
The tarballs produced are different after every build due to meta-data
differences (time) etc.. This may cause problems if openSUSE Build
Service rebuilds a package due to a dependency changing which will cause
the md5sums to change. The package may then fail to upgrade. The
solution is to create reproducible archives so that if the sources do
not change then the archive produced in the build is identical.

Link: https://www.gnu.org/software/tar/manual/html_section/Reproducibility.html
Signed-off-by: Mark Grant <m.grant.prg@gmail.com>